move user and post query tests to gqlgen package
It seems to me that because these tests do not exercise http, they only test gqlgen code.
