+++ application/actions/gql_test.go
@@ -7,14 +7,6 @@ import (
 	"io/ioutil"
 	"strings"
 
-	"github.com/gobuffalo/nulls"
-	"github.com/silinternational/wecarry-api/domain"
-
-	"github.com/silinternational/wecarry-api/gqlgen"
-
-	"github.com/99designs/gqlgen/client"
-	"github.com/99designs/gqlgen/handler"
-
 	"github.com/gobuffalo/httptest"
 	"github.com/silinternational/wecarry-api/models"
 )
@@ -25,38 +17,6 @@ type gqlErrorResponse struct {
 		Path    []string `json:"path"`
 	} `json:"errors"`
 	Data interface{} `json:"data"`
-}
-
-func (as *ActionSuite) Test_UserQuery() {
-	t := as.T()
-	models.ResetTables(t, as.DB)
-
-	queryFixtures := Fixtures_UserQuery(as, t)
-	userFixtures := queryFixtures.Users
-
-	c := getGqlClient()
-
-	query := `{user(id: "` + userFixtures[1].Uuid.String() + `") {id nickname photoURL}}`
-
-	var usersResp struct {
-		User struct {
-			ID       string `json:"id"`
-			Nickname string `json:"nickname"`
-			PhotoURL string `json:"photoURL"`
-		} `json:"user"`
-	}
-
-	gqlgen.TestUser = userFixtures[0]
-	gqlgen.TestUser.AdminRole = nulls.NewString(domain.AdminRoleSuperDuperAdmin)
-	c.MustPost(query, &usersResp)
-
-	if err := as.DB.Load(&(userFixtures[1]), "PhotoFile"); err != nil {
-		t.Errorf("failed to load user fixture, %s", err)
-	}
-	as.Equal(userFixtures[1].Uuid.String(), usersResp.User.ID)
-	as.Equal(userFixtures[1].Nickname, usersResp.User.Nickname)
-	as.Equal(userFixtures[1].PhotoFile.URL.String, usersResp.User.PhotoURL)
-	as.Regexp("^https?", usersResp.User.PhotoURL)
 }
 
 func (as *ActionSuite) Test_CreateOrganization() {
@@ -222,48 +182,4 @@ func (as *ActionSuite) Test_CreateOrganization() {
 		}
 
 	}
-}
-
-func (as *ActionSuite) Test_PostQuery() {
-	t := as.T()
-	models.ResetTables(t, as.DB)
-
-	queryFixtures := Fixtures_PostQuery(as, t)
-	userFixtures := queryFixtures.Users
-	postFixtures := queryFixtures.Posts
-
-	c := getGqlClient()
-
-	query := `{ post(id: "` + postFixtures[1].Uuid.String() + `") { id photo { id } files { id } }}`
-
-	var postsResp struct {
-		Post struct {
-			ID    string `json:"id"`
-			Photo struct {
-				ID string `json:"id"`
-			} `json:"photo"`
-			Files []struct {
-				ID string `json:"id"`
-			} `json:"files"`
-		} `json:"post"`
-	}
-
-	gqlgen.TestUser = userFixtures[0]
-	c.MustPost(query, &postsResp)
-
-	if err := as.DB.Load(&(postFixtures[1]), "PhotoFile", "Files"); err != nil {
-		t.Errorf("failed to load post fixture, %s", err)
-	}
-
-	as.Equal(postFixtures[1].Uuid.String(), postsResp.Post.ID)
-	as.Equal(postFixtures[1].PhotoFile.UUID.String(), postsResp.Post.Photo.ID)
-	as.Equal(1, len(postsResp.Post.Files))
-	as.Equal(postFixtures[1].Files[0].File.UUID.String(), postsResp.Post.Files[0].ID)
-}
-
-func getGqlClient() *client.Client {
-	h := handler.GraphQL(gqlgen.NewExecutableSchema(gqlgen.Config{Resolvers: &gqlgen.Resolver{}}))
-	srv := httptest.NewServer(h)
-	c := client.New(srv.URL)
-	return c
 }
